Big commit (sorry), changed how builds and dependency tracking
