+++ tests/contracts.hpp.in
@@ -4,18 +4,18 @@
 namespace eosio { namespace testing {
 
 struct contracts {
-   static std::vector<uint8_t> system_wasm() { return read_wasm("${ROOT_DIR}/eosio.system/bin/eosio.system/eosio.system.wasm"); }
-   static std::string          system_wast() { return read_wast("${ROOT_DIR}/eosio.system/bin/eosio.system/eosio.system.wast"); }
-   static std::vector<char>    system_abi() { return read_abi("${ROOT_DIR}/eosio.system/bin/eosio.system/eosio.system.abi"); }
-   static std::vector<uint8_t> token_wasm() { return read_wasm("${ROOT_DIR}/eosio.token/bin/eosio.token/eosio.token.wasm"); }
-   static std::string          token_wast() { return read_wast("${ROOT_DIR}/eosio.token/bin/eosio.token/eosio.token.wast"); }
-   static std::vector<char>    token_abi() { return read_abi("${ROOT_DIR}/eosio.token/bin/eosio.token/eosio.token.abi"); } 
-   static std::vector<uint8_t> msig_wasm() { return read_wasm("${ROOT_DIR}/eosio.msig/bin/eosio.msig/eosio.msig.wasm"); }
-   static std::string          msig_wast() { return read_wast("${ROOT_DIR}/eosio.msig/bin/eosio.msig/eosio.msig.wast"); }
-   static std::vector<char>    msig_abi() { return read_abi("${ROOT_DIR}/eosio.msig/bin/eosio.msig/eosio.msig.abi"); }
-   static std::vector<uint8_t> sudo_wasm() { return read_wasm("${ROOT_DIR}/eosio.sudo/bin/eosio.sudo/eosio.sudo.wasm"); }
-   static std::string          sudo_wast() { return read_wast("${ROOT_DIR}/eosio.sudo/bin/eosio.sudo/eosio.sudo.wast"); }
-   static std::vector<char>    sudo_abi() { return read_abi("${ROOT_DIR}/eosio.sudo/bin/eosio.sudo/eosio.sudo.abi"); }
+   static std::vector<uint8_t> system_wasm() { return read_wasm("${CMAKE_BINARY_DIR}/../eosio.system/eosio.system.wasm"); }
+   static std::string          system_wast() { return read_wast("${CMAKE_BINARY_DIR}/../eosio.system/eosio.system.wast"); }
+   static std::vector<char>    system_abi() { return read_abi("${CMAKE_BINARY_DIR}/../eosio.system/eosio.system.abi"); }
+   static std::vector<uint8_t> token_wasm() { return read_wasm("${CMAKE_BINARY_DIR}/../eosio.token/eosio.token.wasm"); }
+   static std::string          token_wast() { return read_wast("${CMAKE_BINARY_DIR}/../eosio.token/eosio.token.wast"); }
+   static std::vector<char>    token_abi() { return read_abi("${CMAKE_BINARY_DIR}/../eosio.token/eosio.token.abi"); } 
+   static std::vector<uint8_t> msig_wasm() { return read_wasm("${CMAKE_BINARY_DIR}/../eosio.msig/eosio.msig.wasm"); }
+   static std::string          msig_wast() { return read_wast("${CMAKE_BINARY_DIR}/../eosio.msig/eosio.msig.wast"); }
+   static std::vector<char>    msig_abi() { return read_abi("${CMAKE_BINARY_DIR}/../eosio.msig/eosio.msig.abi"); }
+   static std::vector<uint8_t> sudo_wasm() { return read_wasm("${CMAKE_BINARY_DIR}/../eosio.sudo/eosio.sudo.wasm"); }
+   static std::string          sudo_wast() { return read_wast("${CMAKE_BINARY_DIR}/../eosio.sudo/eosio.sudo.wast"); }
+   static std::vector<char>    sudo_abi() { return read_abi("${CMAKE_BINARY_DIR}/../eosio.sudo/eosio.sudo.abi"); }
    
    struct util {
       static std::vector<uint8_t> test_api_wasm() { return read_wasm("${CMAKE_SOURCE_DIR}/test_contracts/test_api.wasm"); }
